proper use of size_t instead of int
git-svn-id: http://beaengine.googlecode.com/svn/trunk@21 e67b0f8a-f93e-11de-ba51-6b4196a2b81e

+++ beaengineSources/BeaEngine.h
@@ -290,6 +290,6 @@ enum SPECIAL_INFO
 #ifdef __cplusplus
 extern "C"
 #endif
-int __bea_callspec__ Disasm(LPDISASM pDisAsm);
+Int64 __bea_callspec__ Disasm(LPDISASM pDisAsm);
 
 #endif
